Warning removal

+++ include/boost/test/impl/test_tools.ipp
@@ -733,7 +733,7 @@ output_test_stream::match_pattern( bool flush_stream )
                     for( std::size_t k = 1; k < (std::max)(best_pattern_start_index, best_stream_start_index); k++ ) { // 1 is for the current char c
                         std::string s1(pretty_print_log(std::string(1, last_elements_ordered[(std::min)(k, best_pattern_start_index)])));
                         std::string s2(pretty_print_log(std::string(1, sub_str_suffix[(std::min)(k, best_stream_start_index)])));
-                        for( int h = (std::max)(s1.size(), s2.size()); h > 0; h--)
+                        for( int h = static_cast<int>((std::max)(s1.size(), s2.size())); h > 0; h--)
                             result.message() << "~";
                     }
